refactor: share schema generation
