+++ src/Command/GenerateCommand.php
@@ -4,20 +4,15 @@ declare(strict_types=1);
 
 namespace Helmich\Schema2Class\Command;
 
-use Helmich\Schema2Class\Generator\Property\NestedObjectProperty;
-use Helmich\Schema2Class\Generator\DefinitionsReferenceLookup;
-use Helmich\Schema2Class\Generator\GeneratorException;
 use Helmich\Schema2Class\Generator\GeneratorRequest;
 use Helmich\Schema2Class\Generator\NamespaceInferrer;
-use Helmich\Schema2Class\Generator\SchemaToClass;
 use Helmich\Schema2Class\Generator\SchemaToClassFactory;
 use Helmich\Schema2Class\Loader\LoadingException;
 use Helmich\Schema2Class\Loader\SchemaLoader;
 use Helmich\Schema2Class\Spec\SpecificationFilesItem;
 use Helmich\Schema2Class\Spec\SpecificationOptions;
 use Helmich\Schema2Class\Spec\ValidatedSpecificationFilesItem;
-use Helmich\Schema2Class\Writer\DebugWriter;
-use Helmich\Schema2Class\Writer\FileWriter;
+use Helmich\Schema2Class\Command\GenerateFromRequestTrait;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -26,10 +21,10 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class GenerateCommand extends Command
 {
+    use GenerateFromRequestTrait;
+
     private SchemaLoader $loader;
-
     private NamespaceInferrer $namespaceInferrer;
-
     private SchemaToClassFactory $s2c;
 
     public function __construct(SchemaLoader $loader, NamespaceInferrer $namespaceInferrer, SchemaToClassFactory $s2c)
@@ -79,25 +74,9 @@ class GenerateCommand extends Command
         $output->writeln("loading schema from <comment>$schemaFile</comment>");
         $schema = $this->loader->loadSchema($schemaFile);
 
-        if (!$targetNamespace) {
-            $output->writeln("target namespace not given. inferring from target directory…");
-            try {
-                $targetNamespace = $this->namespaceInferrer
-                    ->inferNamespaceFromTargetDirectory($targetDirectory);
-            } catch (GeneratorException $e) {
-                $output->writeln(
-                    "  ↳ PSR‑4 lookup failed, defaulting to class name as namespace: <comment>{$class}</comment>"
-                );
-                $targetNamespace = $class;
-            }
-        }
+        $targetNamespace = $this->inferNamespace($output, $targetNamespace, $targetDirectory, $class);
 
         $output->writeln("using target namespace <comment>$targetNamespace</comment> in directory <comment>$targetDirectory</comment>");
-
-        $writer = new FileWriter($output);
-        if ($input->getOption("dry-run")) {
-            $writer = new DebugWriter($output);
-        }
 
         $spec = new ValidatedSpecificationFilesItem($targetNamespace, $class, $targetDirectory);
         $opts = (new SpecificationOptions())
@@ -107,27 +86,9 @@ class GenerateCommand extends Command
             $opts = $opts->withTargetPHPVersion("5.6.0");
         }
 
-        // If the schema defines "definitions", emit one class per definition first:
-        $definitions = $schema['definitions'] ?? [];
-        // Build a Lookup so all $ref → definition classes will resolve
-        $lookup = new DefinitionsReferenceLookup($definitions);
-        // Base request with lookup wired in
-        $baseRequest = (new GeneratorRequest($schema, $spec, $opts))
-            ->withReferenceLookup($lookup);
-        // Generate one class per definition
-        foreach ($definitions as $defName => $defSchema) {
-            $reqDef = $baseRequest
-                ->withClass($defName)
-                ->withSchema($defSchema);
-            $this->s2c->build($writer, $output)
-               ->schemaToClass($reqDef);
-        }
-        // Only generate the "main" class if the root schema really defines an object
-        if (NestedObjectProperty::canHandleSchema($schema)) {
-            $mainRequest = $baseRequest->withClass($class);
-            $this->s2c->build($writer, $output)
-                ->schemaToClass($mainRequest);
-        }
+        $baseRequest = new GeneratorRequest($schema, $spec, $opts);
+
+        $this->generateFromRequest($baseRequest, $output, (bool)$input->getOption('dry-run'));
 
         return 0;
     }
